requirements.txt
fastapi==0.100.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
uvicorn==0.23.2
python-multipart